refactor : 재시도 가능한 예외 yml로 추출 및 이벤트 도메인으로 변경
