+++ core/domain/src/main/java/com/loopers/core/domain/order/vo/OrderCompletedEvent.java
@@ -1,0 +1,4 @@
+package com.loopers.core.domain.order.vo;
+
+public record OrderCompletedEvent(OrderId orderId) {
+}
