+++ core/domain/src/main/java/com/loopers/core/domain/payment/vo/PaymentCompletedEvent.java
@@ -1,0 +1,4 @@
+package com.loopers.core.domain.payment.vo;
+
+public record PaymentCompletedEvent(PaymentId paymentId) {
+}
