+++ core/domain/src/main/java/com/loopers/core/domain/payment/vo/PaymentCreatedEvent.java
@@ -1,7 +1,6 @@
-package com.loopers.core.service.payment.event;
+package com.loopers.core.domain.payment.vo;
 
 import com.loopers.core.domain.payment.type.PaymentType;
-import com.loopers.core.domain.payment.vo.PaymentId;
 
 public record PaymentCreatedEvent(PaymentId paymentId, PaymentType paymentType) {
