+++ core/domain/src/main/java/com/loopers/core/domain/payment/vo/PaymentFailedEvent.java
@@ -1,0 +1,4 @@
+package com.loopers.core.domain.payment.vo;
+
+public record PaymentFailedEvent(PaymentId paymentId, FailedReason failedReason) {
+}
