+++ core/service/src/main/java/com/loopers/core/service/order/OrderDataPlatformHandler.java
@@ -3,8 +3,8 @@ package com.loopers.core.service.order;
 import com.loopers.core.domain.order.Order;
 import com.loopers.core.domain.order.OrderDataPlatformClient;
 import com.loopers.core.domain.order.repository.OrderRepository;
+import com.loopers.core.domain.order.vo.OrderCompletedEvent;
 import com.loopers.core.service.order.component.OrderDataPlatformSendingFailHandler;
-import com.loopers.core.service.order.event.OrderCompletedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
