+++ core/service/src/main/java/com/loopers/core/service/order/OrderService.java
@@ -4,6 +4,7 @@ import com.loopers.core.domain.order.Order;
 import com.loopers.core.domain.order.OrderItem;
 import com.loopers.core.domain.order.repository.OrderItemRepository;
 import com.loopers.core.domain.order.repository.OrderRepository;
+import com.loopers.core.domain.order.vo.OrderCompletedEvent;
 import com.loopers.core.domain.order.vo.Quantity;
 import com.loopers.core.domain.product.Product;
 import com.loopers.core.domain.product.repository.ProductRepository;
@@ -12,7 +13,6 @@ import com.loopers.core.domain.user.User;
 import com.loopers.core.domain.user.repository.UserRepository;
 import com.loopers.core.domain.user.vo.UserIdentifier;
 import com.loopers.core.service.order.command.OrderProductsCommand;
-import com.loopers.core.service.order.event.OrderCompletedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
