+++ core/service/src/main/java/com/loopers/core/service/order/PayedOrderHandler.java
@@ -4,8 +4,8 @@ import com.loopers.core.domain.order.Order;
 import com.loopers.core.domain.order.repository.OrderRepository;
 import com.loopers.core.domain.payment.Payment;
 import com.loopers.core.domain.payment.repository.PaymentRepository;
+import com.loopers.core.domain.payment.vo.PaymentCompletedEvent;
 import com.loopers.core.service.order.component.PayedOrderFailHandler;
-import com.loopers.core.service.payment.event.PaymentCompletedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
