+++ core/service/src/main/java/com/loopers/core/service/order/component/PayedOrderFailHandler.java
@@ -1,7 +1,6 @@
 package com.loopers.core.service.order.component;
 
 import com.loopers.JacksonUtil;
-import com.loopers.core.domain.error.HttpClientException;
 import com.loopers.core.domain.event.EventOutbox;
 import com.loopers.core.domain.event.repository.EventOutboxRepository;
 import com.loopers.core.domain.event.type.AggregateType;
@@ -9,35 +8,27 @@ import com.loopers.core.domain.event.type.EventType;
 import com.loopers.core.domain.event.vo.EventPayload;
 import com.loopers.core.domain.order.Order;
 import com.loopers.core.domain.order.repository.OrderRepository;
+import com.loopers.core.domain.order.vo.PayedOrderFailEvent;
 import com.loopers.core.domain.payment.Payment;
 import com.loopers.core.domain.payment.repository.PaymentRepository;
-import com.loopers.core.service.order.event.PayedOrderFailEvent;
-import com.loopers.core.service.payment.event.PaymentCompletedEvent;
+import com.loopers.core.domain.payment.vo.PaymentCompletedEvent;
+import com.loopers.core.service.config.RetryableExceptionsProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-
-import java.util.List;
 
 @Component
 @RequiredArgsConstructor
 public class PayedOrderFailHandler {
 
-    private static final List<Class<? extends Exception>> RETRYABLE_EXCEPTIONS = List.of(
-            HttpClientException.ServiceUnavailable.class,
-            HttpClientException.InternalServerError.class,
-            HttpClientException.GatewayTimeout.class,
-            HttpClientException.TooManyRequests.class
-    );
-
     private final EventOutboxRepository eventOutboxRepository;
     private final OrderRepository orderRepository;
     private final PaymentRepository paymentRepository;
+    private final RetryableExceptionsProperties retryableExceptionsProperties;
 
     public void handle(PaymentCompletedEvent event, Exception exception) {
         Payment payment = paymentRepository.getById(event.paymentId());
         Order order = orderRepository.getBy(payment.getOrderKey());
-        boolean retryable = RETRYABLE_EXCEPTIONS.stream()
-                .anyMatch(exceptionClass -> exceptionClass.isInstance(exception));
+        boolean retryable = retryableExceptionsProperties.isRetryable(exception);
 
         eventOutboxRepository.save(
                 EventOutbox.create(
