--- core/service/src/main/java/com/loopers/core/service/order/event/OrderCompletedEvent.java
@@ -1,6 +1,0 @@
-package com.loopers.core.service.order.event;
-
-import com.loopers.core.domain.order.vo.OrderId;
-
-public record OrderCompletedEvent(OrderId orderId) {
-}
