+++ core/service/src/main/java/com/loopers/core/service/payment/PaymentDataPlatformEventHandler.java
@@ -3,8 +3,8 @@ package com.loopers.core.service.payment;
 import com.loopers.core.domain.payment.Payment;
 import com.loopers.core.domain.payment.PaymentDataPlatformClient;
 import com.loopers.core.domain.payment.repository.PaymentRepository;
+import com.loopers.core.domain.payment.vo.PaymentCompletedEvent;
 import com.loopers.core.service.payment.component.PaymentDataPlatformSendingFailHandler;
-import com.loopers.core.service.payment.event.PaymentCompletedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
