+++ core/service/src/main/java/com/loopers/core/service/payment/PgCallbackService.java
@@ -6,11 +6,11 @@ import com.loopers.core.domain.payment.PgPayment;
 import com.loopers.core.domain.payment.repository.PgPaymentRepository;
 import com.loopers.core.domain.payment.type.PgPaymentStatus;
 import com.loopers.core.domain.payment.vo.FailedReason;
+import com.loopers.core.domain.payment.vo.PaymentCompletedEvent;
 import com.loopers.core.domain.payment.vo.TransactionKey;
 import com.loopers.core.service.payment.command.PgCallbackCommand;
 import com.loopers.core.service.payment.component.PaymentCallbackStrategy;
 import com.loopers.core.service.payment.component.PaymentCallbackStrategySelector;
-import com.loopers.core.service.payment.event.PaymentCompletedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
