+++ core/service/src/main/java/com/loopers/core/service/payment/PgPaymentHandler.java
@@ -7,9 +7,9 @@ import com.loopers.core.domain.payment.repository.PaymentRepository;
 import com.loopers.core.domain.payment.repository.PgPaymentRepository;
 import com.loopers.core.domain.payment.type.PaymentType;
 import com.loopers.core.domain.payment.vo.FailedReason;
+import com.loopers.core.domain.payment.vo.PaymentCreatedEvent;
 import com.loopers.core.domain.payment.vo.PaymentId;
 import com.loopers.core.service.payment.component.PgPaymentRequestFailHandler;
-import com.loopers.core.service.payment.event.PaymentCreatedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
