--- core/service/src/main/java/com/loopers/core/service/payment/event/PaymentCompletedEvent.java
@@ -1,6 +1,0 @@
-package com.loopers.core.service.payment.event;
-
-import com.loopers.core.domain.payment.vo.PaymentId;
-
-public record PaymentCompletedEvent(PaymentId paymentId) {
-}
