--- core/service/src/main/java/com/loopers/core/service/payment/event/PaymentFailedEvent.java
@@ -1,7 +1,0 @@
-package com.loopers.core.service.payment.event;
-
-import com.loopers.core.domain.payment.vo.FailedReason;
-import com.loopers.core.domain.payment.vo.PaymentId;
-
-public record PaymentFailedEvent(PaymentId paymentId, FailedReason failedReason) {
-}
