+++ core/service/src/main/java/com/loopers/core/service/product/StockIncreaseHandler.java
@@ -6,9 +6,9 @@ import com.loopers.core.domain.order.repository.OrderItemRepository;
 import com.loopers.core.domain.order.repository.OrderRepository;
 import com.loopers.core.domain.payment.Payment;
 import com.loopers.core.domain.payment.repository.PaymentRepository;
+import com.loopers.core.domain.payment.vo.PaymentFailedEvent;
 import com.loopers.core.domain.product.Product;
 import com.loopers.core.domain.product.repository.ProductRepository;
-import com.loopers.core.service.payment.event.PaymentFailedEvent;
 import com.loopers.core.service.product.component.StockIncreaseFailHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
