+++ core/service/src/test/java/com/loopers/core/service/product/StockIncreaseHandlerIntegrationTest.java
@@ -10,6 +10,7 @@ import com.loopers.core.domain.payment.Payment;
 import com.loopers.core.domain.payment.PaymentFixture;
 import com.loopers.core.domain.payment.repository.PaymentRepository;
 import com.loopers.core.domain.payment.vo.FailedReason;
+import com.loopers.core.domain.payment.vo.PaymentFailedEvent;
 import com.loopers.core.domain.product.Product;
 import com.loopers.core.domain.product.ProductFixture;
 import com.loopers.core.domain.product.repository.ProductRepository;
@@ -21,7 +22,6 @@ import com.loopers.core.domain.user.UserFixture;
 import com.loopers.core.domain.user.repository.UserRepository;
 import com.loopers.core.service.ConcurrencyTestUtil;
 import com.loopers.core.service.IntegrationTest;
-import com.loopers.core.service.payment.event.PaymentFailedEvent;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
